requirements.txt
streamlit
tensorflow
pillow
huggingface_hub
numpy
python-dotenv

